Fix data type

+++ src/db/query/ProjectDonatorQuery.java
@@ -79,7 +79,7 @@ public class ProjectDonatorQuery {
             try {
     			while (resultSet.next()) {
     				final Integer idProgetto = resultSet.getInt("idProgetto");
-    				final Integer donazioniProgetto = resultSet.getInt("donazioniProgetto");
+    				final BigDecimal donazioniProgetto = resultSet.getBigDecimal("donazioniProgetto");
     				final String percentuale = resultSet.getString("percentuale");
 
 					Object[] data = { idProgetto, donazioniProgetto, percentuale };
